LPS-96587 Not needed

+++ modules/apps/change-tracking/change-tracking-rest-impl/src/main/java/com/liferay/change/tracking/rest/internal/resource/v1_0/AffectedEntryResourceImpl.java
@@ -73,7 +73,7 @@ public class AffectedEntryResourceImpl extends BaseAffectedEntryResourceImpl {
 						entryId, keywords, queryDefinition),
 					this::_toAffectedEntry),
 				pagination,
-				(int)_ctEntryLocalService.getRelatedOwnerCTEntriesCount(
+				_ctEntryLocalService.getRelatedOwnerCTEntriesCount(
 					ownerCTEntry.getCompanyId(), collectionId,
 					entryId, keywords, queryDefinition));
 		}
